Regression for Razer Blade Laptops
Fix HID-RMI related regression for Razer Blade laptops. The touchpads on
these laptops are strange, and have caused issues in the past. Commit
0925636042170e0b6716cd86635899c5f4258f69 broke suspend functionality due
to some quirkiness of the touchpad (suspend hangs now).

Found the faulty commit via git bisect, and testing showed that the
razer blade crashes/hangs in rmi_write, and rmi_read_block for an
unknown reason. This solution is temporary, and just disables
rmi_suspend, since it is not working in this case anyway.

Resume also causes suspend to hang, if it is resumed from a closed LID
state, so that needed to be fixed as well.